Refactor hooks and simplify button
